==> 01 Tweet Data Collection.ipynb <==
import pandas as pd
import GetOldTweets3 as got
from time import sleep

def get_tweets(username, top_only, start_date, end_date, max_tweets):

    tweetCriteria = got.manager.TweetCriteria().setUsername(username)\
                    .setTopTweets(top_only)\
                    .setSince(start_date)\
                    .setUntil(end_date)\
                    .setMaxTweets(max_tweets)

    tweets = got.manager.TweetManager.getTweets(tweetCriteria)
    print(f"Total tweets scraped: {len(tweets)}")

    text_tweets = [[tw.username,
                tw.text,
                tw.date,
                tw.retweets,
                tw.favorites,
                tw.mentions,
                tw.hashtags] for tw in tweets]
    
    news_df = pd.DataFrame(text_tweets, 
                            columns = ['User', 'Text', 'Date', 'Favorites', 'Retweets', 'Mentions', 'HashTags'])
    
    return news_df

news_sources = ['La_DOTD','TxDot','BR_Traffic', 'NOLA_Traffic','Laf_Traffic', 'NS_Traffic', 
                'LC_Traffic', 'wbrztraffic', 'WAFBTraffic', 'LAStatePolice', 'BTRtraffic']

news_df = get_tweets(news_sources, 
                     top_only = False,
                     start_date = "2020-08-07", 
                     end_date = "2020-09-04",
                     max_tweets = 2000).sort_values('Date', ascending=False)
news_df.head()

news_df.shape

news_df.head(50)

news_df.to_csv('tweets_data/news_df.csv', index = False)

def get_tweets2(keyword_list, top_only, city, start_date, end_date, max_tweets):
    
    df = pd.DataFrame()
    
    for i in keyword_list:

        tweetCriteria = got.manager.TweetCriteria().setQuerySearch(i)\
                    .setTopTweets(top_only)\
                    .setNear(city)\
                    .setWithin("200mi")\
                    .setSince(start_date)\
                    .setUntil(end_date)\
                    .setMaxTweets(max_tweets)

        tweets = got.manager.TweetManager.getTweets(tweetCriteria)

        text_tweets = [[tw.username,
                tw.text,
                tw.date,
                tw.retweets,
                tw.favorites,
                tw.mentions,
                tw.hashtags] for tw in tweets]
    
        df2 = pd.DataFrame(text_tweets, 
                            columns = ['User', 'Text', 'Date', 'Favorites', 'Retweets', 'Mentions', 'HashTags'])
        
        df = df.append(df2)
        
        print(f"Total tweets scraped: {len(df)}")
        
        sleep(300)
        
    return df
    

keywords1 = ['hurricane', 'closed', 'close']
keywords2 = ['congestion', 'blocked', 'highway', 
            'traffic']
keywords3 = ['route', 'open', 'lane', 'road']
keywords4 = ['shut down', 'detour', 'evacuation', 'hwy']

news4_df = get_tweets2(keyword_list = keywords4,
                     top_only = False,
                     city = "Lake Charles, Louisiana",
                     start_date = "2020-08-24", 
                     end_date = "2020-09-01",
                     max_tweets = 10_000).sort_values('Date', ascending=False)
    
news4_df.head()

news4_df.shape

news4_df.tail()

news4_df.to_csv('tweets_data/news4_df.csv', index = False)

csv_list = ["news_df.csv", "news1_df.csv", "news2_df.csv", "news3_df.csv", "news4_df.csv"]

df_list = []
for i in csv_list:
    df_list.append(pd.read_csv('tweets_data/' + i))

lake_ch_tweets_2 = pd.concat(df_list)

lake_ch_tweets_2.shape

lake_ch_tweets_2.to_csv('tweets_data/lake_ch_tweets_2.csv', index = False)
