==> 01.1 LADOT Data Collection.ipynb <==
import numpy as np
import pandas as pd
import requests
from bs4 import BeautifulSoup

Link to LA DoTD traffic updates

url = 'http://wwwapps.dotd.la.gov/administration/announcements/home.aspx?type=roadandlaneclosure'
print(url)

Connect with LA DoTD

req = requests.get(url)
req.status_code

Instantiate soup object from LA DoTD

soup = BeautifulSoup(req.content, 'lxml')

Pull out updates table from soup

table = soup.find('table', attrs={'id': 'gvAnnouncements'})

Pull reports and dates from every update available

reports = []

for row in table.find_all('tr')[1:]:#5001]:
    info = row.text.strip().split('\n')
    
    report = {}
    report['date'] = info[0]
    report['report'] = info[1]
    
    reports.append(report)
    
reports = pd.DataFrame(reports)
reports.tail()    

reports.shape

Save dataframe of updates

reports.to_csv('./LADOTD Data/road_closures.csv', header=True, index=False)

reports['report'].head(10).values
