==> 02 Tweet NLP.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

import spacy
from spacy.tokens import Span
from spacy.matcher import Matcher
from spacy.pipeline import EntityRuler

# Contents

- [Function Definitions](#fdef)
- [NLP Set-up](#NLP)
- [Data Processing](#data)

# Function Definitions
<a id = "fdef"></a>

---------

# processes a dataframe with a 'Text' row using spacy nlp
# returns a copy of the dataframe with a new 'docs' column

def get_docs(df, nlp):
    new_df = df.copy()
    len_df = len(df)
    docs = []
    
    count = 0
    
    for text in df['Text']:
        try:
            doc = nlp(text.lower())
        except:
            doc = []
        
        count += 1
        if count % 1000 == 0:
            print(f'nlp progress: {round(100.0 * count / len_df, 1)}%')
        
        docs.append(doc)
    
    new_df['docs'] = docs
        
    return new_df

def find_matches(df):
    new_df = df.copy()
    
    to_keep = []
    
    count = 0
    
    row_counter = 0

    for doc in new_df['docs']:
        
        if doc == []:
            to_keep.append(False)
            continue
        
        keep = False

        for ent in doc.ents:
            e = ent.label_
            if(e == "street" or e == 'highway' or e == 'exit'):
                keep = True
                row_counter+=1
                continue

        to_keep.append(keep)
        
        count += 1
    
    print(f"Tried to find matches in {count} documents, found {row_counter} rows with matches.")
    
    return new_df[to_keep]

#This function finds where the docs have our new entity types and creates new dataframe rows
#that correspond to each entity type

def make_street_cols(df):
    new_df = df.copy()
    
    streets = []
    highways = []
    exits = []
    markers = []
    
    for doc in df['docs']:
        doc_streets = []
        doc_highways = []
        doc_exits = []
        doc_markers = []
        
        if(doc != []):
            for ent in doc.ents:
                if(ent.label_ == 'street'):
                    doc_streets.append(ent.text)
                elif(ent.label_ == 'highway'):
                    doc_highways.append(ent.text)
                elif(ent.label_ == 'exit'):
                    doc_exits.append(ent.text)
                elif(ent.label_ == 'marker'):
                    doc_markers.append(ent.text)
        
        streets.append(doc_streets)
        highways.append(doc_highways)
        exits.append(doc_exits)
        markers.append(doc_markers)
    
    new_df['streets'] = streets
    new_df['highways'] = highways
    new_df['exits'] = exits
    new_df['markers'] = markers
    
    return new_df

def extract_road_info(df, nlp):
    new_df = df.drop_duplicates().copy()
    new_df = get_docs(new_df, nlp)
    new_df = find_matches(new_df)
    new_df = make_street_cols(new_df)
    return new_df

# NLP
------

nlp = spacy.load('en_core_web_sm')

#Patterns for entity ruler

street_suff = ['street', 'st', 'avenue', 'ave', 'boulevard', 'blvd', 'highway', 'hwy',
              'circle', 'drive', 'lane', 'road', 'rd', 'way', 'place', 'court',
              'parkway', 'pkwy', 'turnpike', 'tpke', 'trnpk', 'turnpk']

street_pattern = {
    'label' : "street",
    'pattern': [
        {"POS" : {"IN" : ["NOUN","PROPN", "ADJ"]}},
        {"LOWER" : {"IN" : street_suff }} ]
}

#matches strings that look like interstates, ie. I-5, I-15, I-345
interstate_pattern = {
    'label' : "highway",
    'pattern' : [
        {"TEXT" : {"REGEX" : "[Ii]-\d*"}} ]
}

#matches strings that look like US highways, ie. US-123
us_hwy_pattern = {    
    'label' : "highway",
    'pattern' : [
        {"TEXT" : {"REGEX" : "[Uu][Ss]-\d*"}} ]
}

#matches strings that look like other highway names, including 'US' without the dash
oth_hwy_pattern = {    
    'label' : "highway",
    'pattern' : [
        {'LOWER' : {"IN" : ['us', 'highway', 'hwy', 'route', 'rt', 'rte']}},
        {"LIKE_NUM" : True} ]
}

#matches spans that look like exit ramp numbers
exit_pattern = {
    'label' : "exit",
    'pattern' : [
        {"LOWER" : "exit"},
        {"LIKE_NUM" : True} ]
}

state_hwy_pattern = {
    'label' : "highway",
    'pattern' : [
        {"LOWER" : {"IN" : ['la', 'fl', 'al', 'ga', 'nc', 'sc', 'ms']}},
        {"LIKE_NUM" : True}
    ]
}

#matches spans that look like mile markers
marker_pattern = {
    'label' : "marker",
    'pattern' : [
        {"LOWER" : "mile"},
        {"LOWER" : {"IN" : ["marker", "post"]}}, {"LIKE_NUM" : True} ]
}

patterns = [street_pattern, interstate_pattern, us_hwy_pattern, state_hwy_pattern, oth_hwy_pattern, exit_pattern, marker_pattern]

ruler = EntityRuler(nlp)
ruler.add_patterns(patterns)
nlp.add_pipe(ruler, before = 'ner')

# Data Processing
<a id="data"></a>

-----

tweets_df = pd.read_csv('./tweets_data/lake_ch_tweets.csv')

tweets_df.info()

%%time
tweets_df_streets = extract_road_info(tweets_df, nlp)

tweets_df_streets.head()

tweets_df_streets.info()

tweets_df_streets.to_csv('tweets_data/tweets_df_extended.csv')

nlp.pipeline

import pickle

pickle.dump(nlp, open('./models/tweet_nlp.p', 'wb'))

pickle.dump(rf, open('./models/RoadFinder.p', 'wb'))
