==> 04 GeoCode with Here API.ipynb <==
import requests
import pandas as pd
import numpy as np
import time as time

import seaborn as sns
import matplotlib.pyplot as plt


#allow all columns to be seen. See show all columns
pd.set_option('display.max_columns', None)

# Source:https://stackoverflow.com/questions/47022070/display-all-dataframe-columns-in-a-jupyter-python-notebook/47022213 

#all text to be seen
pd.set_option('display.max_colwidth', None)

#https://stackoverflow.com/questions/25351968/how-to-display-full-non-truncated-dataframe-information-in-html-when-convertin 

# Read in the Twitter Data

tweets = pd.read_csv('./tweets_data/tweets_predictions.csv')
tweets.head(2)

tweets.shape

# Prep the Data

## Convert the data into a `query_string` column

#Take streets, highways, exits, and markers and put them in one column
tweets['query'] = tweets['streets'] + tweets['highways'] + tweets['exits'] + tweets['markers']

#remove brackets
tweets['query'] = [val.replace('[]','') for val in tweets['query'].values]
tweets['query'] = [val.replace('][',',') for val in tweets['query'].values]
tweets['query'] = [val.replace(']','') for val in tweets['query'].values]
tweets['query'] = [val.replace('[','') for val in tweets['query'].values]

query = []

#url encode the column
for val in tweets['query']:
    val = val.replace(',','%2C+')
    val = val.replace(' ','+')
    val = val.replace("'",'')
    val = val.replace("",'')
    s = ''.join(val)
    query.append(s)

#create the query_string column
tweets['query_string'] = query
tweets.head(2)

# GET Lat/Lng from HERE API

`queryScore` – a value from `0` to `1` representing the percentage of the input that matched the returned address. It is equal to `1` if all input tokens were recognized and matched. If the input query contains any additional information – name of the person, phone number or a hint to ring twice – the `queryScore` will be low, but the result may be correct. But low queryScore may indicate that the result does not match some relevant part of the address in the input query.

`fieldScore` – values from `0` to `1` indicating how good the result field matches to the corresponding part of the query. If all the fields that are important for addressing in the country are matched with high quality, one can assume, that queryScore is low because of the additional non-address information. And the result is still confident.

Resource: https://developer.here.com/documentation/geocoding-search-api/dev_guide/topics-api/code-geocode-hybrid.html

api_key = '<add api key here>' #commented out for privacy reasons

# get the lat and long for all values

#there must be a 'query_string' column in the data frame
def get_lat_lng(df,state):
    lat = []
    lng = []
    query_score = []
    field_score = []

    for row in df['query_string']:
        res=requests.get('https://geocode.search.hereapi.com/v1/geocode?q=' + row + '&qq=state='+ state + '&apiKey=' + api_key) #create the api get request
        time.sleep(2) #add a sleep to not hit servers too hard
        json = res.json() #extract json

        #if no value is returned, just put 0.0
        try:
            #append to lat
            lat.append(json['items'][0]['position']['lat'])

            #append to long
            lng.append(json['items'][0]['position']['lng'])

            #append to query_score
            query_score.append(json['items'][0]['scoring']['queryScore'])

            #append to field_score
            field_score.append(json['items'][0]['scoring']['fieldScore'])

        except IndexError:
            #append to lat
            lat.append(0.0)

            #append to long
            lng.append(0.0)

            #append to query_score
            query_score.append(0.0)

            #append to field_score
            field_score.append(0.0)
        
        except KeyError:
            #append to lat
            lat.append(0.0)

            #append to long
            lng.append(0.0)

            #append to query_score
            query_score.append(0.0)

            #append to field_score
            field_score.append(0.0)
    
    #Add columns to the dataframe
    df['lat'] = lat
    df['lng'] = lng
    df['query_score'] = query_score
    df['field_score'] = field_score
    
    return df


# GET lat/lng from Twitter Data

time0 = time.time() #start timer

tweets_with_geo = get_lat_lng(tweets,'Louisana') #run function

print(f'Time to run: {(time.time() - time0) / 60} mins')

tweets.to_csv('./GeoCoding Data/tweets_with_geo.csv',index=False)

# LADOT FEED

ladot_feed = pd.read_csv('./LADOTD Data/road_closures.csv')
ladot_feed.head(3)

## Create the query string

ladot_feed['query_string'] = [val.replace(' ','+') for val in ladot_feed['report']] #here api will process out the extra words
ladot_feed.head(3) #check work

## Run the function to get lat/lng

time0 = time.time()

ladot_feed_with_geo =  get_lat_lng(ladot_feed,'Louisiana')

print(f'Time to run: {(time.time() - time0) / 60} mins')

ladot_feed_with_geo.to_csv('./GeoCoding Data/ladot_feed_with_geo.csv',index=False)
